Rename Models.Dynamic namespace to Models.GameObjects
