+++ GearBox.Core.Tests/Model/GameObjects/Player/PlayerCharacterTester.cs
@@ -1,8 +1,8 @@
-using GearBox.Core.Model.Dynamic.Player;
+using GearBox.Core.Model.GameObjects.Player;
 using GearBox.Core.Model.Stable.Items;
 using Xunit;
 
-namespace GearBox.Core.Tests.Model.Dynamic.Player;
+namespace GearBox.Core.Tests.Model.GameObjects.Player;
 
 public class PlayerCharacterTester
 {
