+++ GearBox.Core.Tests/Model/Stable/Items/LootChestTester.cs
@@ -1,4 +1,4 @@
-using GearBox.Core.Model.Dynamic.Player;
+using GearBox.Core.Model.GameObjects.Player;
 using GearBox.Core.Model.Stable.Items;
 using GearBox.Core.Model.Units;
 using Xunit;
