+++ GearBox.Core.Tests/Server/SpyControlCommand.cs
@@ -1,6 +1,6 @@
 using GearBox.Core.Controls;
 using GearBox.Core.Model;
-using GearBox.Core.Model.Dynamic.Player;
+using GearBox.Core.Model.GameObjects.Player;
 
 namespace GearBox.Core.Tests.Server;
