+++ GearBox.Core/Model/GameObjects/Player/IPlayerStat.cs
@@ -1,4 +1,4 @@
-namespace GearBox.Core.Model.Dynamic.Player;
+namespace GearBox.Core.Model.GameObjects.Player;
 
 // allows PlayerStatType to get around issues with generics: https://stackoverflow.com/a/15575106
 public interface IPlayerStat
