+++ GearBox.Core/Model/GameObjects/Player/PlayerStat.cs
@@ -1,4 +1,4 @@
-namespace GearBox.Core.Model.Dynamic.Player;
+namespace GearBox.Core.Model.GameObjects.Player;
 
 public class PlayerStat<T> : IPlayerStat
 {
