+++ GearBox.Core/Model/GameObjects/Player/PlayerStatBoosts.cs
@@ -1,4 +1,4 @@
-namespace GearBox.Core.Model.Dynamic.Player;
+namespace GearBox.Core.Model.GameObjects.Player;
 
 public class PlayerStatBoosts
 {
