+++ GearBox.Core/Model/GameObjects/Player/PlayerStats.cs
@@ -1,4 +1,4 @@
-namespace GearBox.Core.Model.Dynamic.Player;
+namespace GearBox.Core.Model.GameObjects.Player;
 
 public class PlayerStats
 {
