+++ GearBox.Core/Model/GameObjects/Projectile.cs
@@ -2,7 +2,7 @@ using System.Text.Json;
 using GearBox.Core.Model.Json;
 using GearBox.Core.Model.Units;
 
-namespace GearBox.Core.Model.Dynamic;
+namespace GearBox.Core.Model.GameObjects;
 
 /// <summary>
 /// Projectiles are generated by characters when they use attacks, and can inflict damage
