+++ GearBox.Core/Model/Stable/Items/Equipment.cs
@@ -1,4 +1,4 @@
-using GearBox.Core.Model.Dynamic.Player;
+using GearBox.Core.Model.GameObjects.Player;
 
 namespace GearBox.Core.Model.Stable.Items;
