+++ GearBox.Core/Model/WorldBuilder.cs
@@ -1,5 +1,5 @@
-using GearBox.Core.Model.Dynamic;
-using GearBox.Core.Model.Dynamic.Player;
+using GearBox.Core.Model.GameObjects;
+using GearBox.Core.Model.GameObjects.Player;
 using GearBox.Core.Model.Stable.Items;
 using GearBox.Core.Model.Static;
 using GearBox.Core.Model.Units;
