+++ GearBox.Core/Server/WorldServer.cs
@@ -1,7 +1,7 @@
 using GearBox.Core.Controls;
 using GearBox.Core.Model;
-using GearBox.Core.Model.Dynamic;
-using GearBox.Core.Model.Dynamic.Player;
+using GearBox.Core.Model.GameObjects;
+using GearBox.Core.Model.GameObjects.Player;
 using GearBox.Core.Model.Json;
 using GearBox.Core.Model.Units;
